added 'lastrowandcolumn'
Added a function lastrowcol in main.py to give the row number and column
number after which there is no more content on the worksheet, and the
corresponding implementation in win32com/windows and appscript/osx.
We discussed this a bit on stackoverflow. I started with working from
'xlLastCell' in Windows because that would prevent looking at rows or
columns with only deleted content, but I could not figure out how to do
that in appscript, so I used 'usedrange' there. I read somewhere that
usedrange also includes rows/columns with deleted content though when I
tried, that did not appear to be the case.