make refers_to_range deal with quotes
The address produced with external=True includes quotes around the workbook/sheet pair in the form `'[book.xlsm]sheet'!$A$1`. The current regex results in sheet name that includes the trailing quote, which results in reference to a non-existent sheet.

This patch excludes the trailing quote from the group that is expected to match the sheet name.