display function MySum with typed parameters
I don't know why the typed MySum functions was not displayed within the docs? Perhaps also because a (missing) return? Could a return here display the function correctly. Or was it the ``` ticks where it went wrong?
Either way, I hope it is fixed. 
If not then I need to figure out how to preview the docs before committing.